Refactor: one declaration per className
In jsxstyle 1 and 2, a className can represent one or more style declarations. This lead to a CSS rules with a significant number of duplicate style declarations. In jsxstyle 3, each className corresponds to a single declaration. Downside: className props will be lengthier. Upside: CSS, both injected and extracted, is significantly easier to optimise. On the injection side, this will greatly reduce the number of CSS rules written to the DOM with addRule. On the extraction side, we can dedupe CSS using really simple techniques (e.g. `dedupe(css.split('\n')).join('\n')`) and experience no downsides.